[FIX]: Se elimino todo lo relacionado con TOPT por limitaciones de diseño

+++ Micro_Users/Users_Aplication/Commands/CreateUserCommand.py
@@ -3,10 +3,10 @@ from typing import Optional
 
 
 class CreateUserCommand(BaseModel):
-	password: str
-	email: EmailStr
-	first_name: Optional[str] = None
-	last_name: Optional[str] = None
-	cedula: float
+    password: str
+    email: EmailStr
+    first_name: Optional[str] = None
+    last_name: Optional[str] = None
+    cedula: int
